importing review and user data

+++ dev-data/data/import-dev-data.js
@@ -3,6 +3,8 @@ const mongoose = require('mongoose');
 const dotenv = require('dotenv');
 
 const Tour = require('../../models/tourModel');
+const User = require('../../models/userModel');
+const Review = require('../../models/reviewModel');
 
 dotenv.config({ path: './config.env' });
 
@@ -25,11 +27,18 @@ mongoose
 
 // read JSON file
 const tours = JSON.parse(fs.readFileSync(`${__dirname}/tours.json`, 'utf-8'));
+const users = JSON.parse(fs.readFileSync(`${__dirname}/users.json`, 'utf-8'));
+const reviews = JSON.parse(
+  fs.readFileSync(`${__dirname}/reviews.json`, 'utf-8')
+);
 
 // import data into DB
 const importData = async () => {
   try {
     await Tour.create(tours);
+    await User.create(users);
+    await Review.create(reviews);
+
     console.log('Data successfully loaded to the database!');
   } catch (err) {
     console.log(`Error importing data to the database: ${err}`);
@@ -40,6 +49,9 @@ const importData = async () => {
 const deleteData = async () => {
   try {
     await Tour.deleteMany();
+    await User.deleteMany();
+    await Review.deleteMany();
+
     console.log('Data successfully deleted from the database!');
   } catch (err) {
     console.log(`Error deleting data from the database: ${err}`);
@@ -47,10 +59,5 @@ const deleteData = async () => {
   process.exit();
 };
 
-if (process.argv[2] === '--import') {
-  importData();
-} else if (process.argv[2] === '--delete') {
-  deleteData();
-}
-
-// console.log(process.argv);
+if (process.argv[2] === '--import') importData();
+if (process.argv[2] === '--delete') deleteData();
